OSF release 1.3.5-osf from origin 1.3.5

+++ core/src/main/java/fr/gael/dhus/olingo/v1/entityset/UserSynchronizerEntitySet.java
@@ -139,7 +139,7 @@ public class UserSynchronizerEntitySet extends AbstractEntitySet<UserSynchronize
       );
 
       properties.add(new SimpleProperty().setName(PAGE_SIZE)
-            .setType(EdmSimpleTypeKind.Int64)
+            .setType(EdmSimpleTypeKind.Int32)
             .setFacets(
                   new Facets()
                         .setNullable(false)
